+++ webserver/app/tasks_api.py
@@ -6,12 +6,15 @@ tasks-related endpoints:
 - POST /tasks/validate
 - GET /tasks/id
 - POST /tasks/id/cancel
+- GET /tasks/id/results
+- POST /tasks/id/results/approve
+- POST /tasks/id/results/block
 """
 from datetime import datetime, timedelta
 from flask import Blueprint, request, send_file
 
 from app.helpers.const import CLEANUP_AFTER_DAYS
-from app.helpers.exceptions import DBRecordNotFoundError, UnauthorizedError
+from app.helpers.exceptions import UnauthorizedError, InvalidRequest
 from app.helpers.keycloak import Keycloak
 from app.helpers.wrappers import audit, auth
 from app.helpers.db import db
@@ -51,9 +54,7 @@ def get_task_id(task_id):
     """
     GET /tasks/id endpoint. Gets a single task
     """
-    task = Task.query.filter(Task.id == task_id).one_or_none()
-    if task is None:
-        raise DBRecordNotFoundError(f"Dataset with id {task_id} does not exist")
+    task = Task.get_by_id(task_id)
 
     token = request.headers.get("Authorization", "").replace("Bearer ", "")
     kc_client = Keycloak()
@@ -62,9 +63,7 @@ def get_task_id(task_id):
     if task.requested_by != dec_token['sub'] and not kc_client.is_user_admin(token):
         raise UnauthorizedError("User does not have enough permissions")
 
-    task_dict = task.sanitized_dict()
-    task_dict["status"] = task.get_status()
-    return task_dict, 200
+    return task.sanitized_dict(), 200
 
 @bp.route('/<task_id>/cancel', methods=['POST'])
 @audit
@@ -73,9 +72,7 @@ def cancel_tasks(task_id):
     """
     POST /tasks/id/cancel endpoint. Cancels a task either scheduled or running one
     """
-    task = Task.query.filter(Task.id == task_id).one_or_none()
-    if task is None:
-        raise DBRecordNotFoundError(f"Task with id {task_id} does not exist")
+    task = Task.get_by_id(task_id)
 
     # Should remove pod/stop ML pipeline
     return task.terminate_pod(), 201
@@ -116,14 +113,56 @@ def post_tasks_validate():
 def get_task_results(task_id):
     """
     GET /tasks/id/results endpoint.
-        Allows to get tasks results
+        Allows to get tasks results if approved to be released
+        or, if an admin is trying to view them
     """
-    task = Task.query.filter(Task.id == task_id).one_or_none()
-    if task is None:
-        raise DBRecordNotFoundError(f"Dataset with id {task_id} does not exist")
+    token = request.headers.get("Authorization", "").replace("Bearer ", "")
+    kc_client = Keycloak()
+
+    task = Task.get_by_id(task_id)
+
+    if not task.review_status and not kc_client.is_user_admin(token):
+        return {"status": task.get_review_status()}, 400
 
     if task.created_at.date() + timedelta(days=CLEANUP_AFTER_DAYS) <= datetime.now().date():
         return {"error": "Tasks results are not available anymore. Please, run the task again"}, 500
 
     results_file = task.get_results()
     return send_file(results_file, download_name="results.tar.gz"), 200
+
+@bp.route('/<task_id>/results/approve', methods=['POST'])
+@audit
+@auth(scope='can_admin_task')
+def approve_results(task_id):
+    """
+    POST /tasks/id/results/approve endpoint.
+        Approves the release (automatic or manual) of
+        a task's results
+    """
+    task = Task.get_by_id(task_id)
+    if task.review_status is not None:
+        raise InvalidRequest("Task has been already reviewed")
+    task.review_status = True
+
+    return {
+        "status": task.get_review_status()
+    }, 201
+
+@bp.route('/<task_id>/results/block', methods=['POST'])
+@audit
+@auth(scope='can_admin_task')
+def block_results(task_id):
+    """
+    POST /tasks/id/results/block endpoint.
+        Blocks the release (automatic or manual) of
+        a task's results
+    """
+    task = Task.get_by_id(task_id)
+    if task.review_status is not None:
+        raise InvalidRequest("Task has been already reviewed")
+
+    task.review_status = False
+
+    return {
+        "status": task.get_review_status()
+    }, 201
